SenseiEngine/services/engine/requirements.txt
fastapi
uvicorn
mysql-connector-python
google-generativeai
python-dotenv
